Move SentryConfig to sentry package
