+++ src/main/java/net/wurstclient/hack/Hack.java
@@ -11,9 +11,9 @@ import java.util.Objects;
 
 import net.wurstclient.Category;
 import net.wurstclient.Feature;
-import net.wurstclient.SentryConfig;
 import net.wurstclient.hacks.NavigatorHack;
 import net.wurstclient.hacks.TooManyHaxHack;
+import net.wurstclient.sentry.SentryConfig;
 
 public abstract class Hack extends Feature
 {
