+++ src/main/java/net/wurstclient/mixin/MinecraftClientMixin.java
@@ -29,13 +29,13 @@ import net.minecraft.util.hit.EntityHitResult;
 import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.snooper.SnooperListener;
 import net.minecraft.util.thread.ReentrantThreadExecutor;
-import net.wurstclient.SentryConfig;
 import net.wurstclient.WurstClient;
 import net.wurstclient.events.LeftClickListener.LeftClickEvent;
 import net.wurstclient.events.RightClickListener.RightClickEvent;
 import net.wurstclient.mixinterface.IClientPlayerEntity;
 import net.wurstclient.mixinterface.IClientPlayerInteractionManager;
 import net.wurstclient.mixinterface.IMinecraftClient;
+import net.wurstclient.sentry.SentryConfig;
 
 @Mixin(MinecraftClient.class)
 public abstract class MinecraftClientMixin
